+++ src/RoadCaptain.App.Runner/ViewModels/DesignTimeInGameViewModel.cs
@@ -81,7 +81,8 @@ namespace RoadCaptain.App.Runner.ViewModels
                 }, 
                 DefaultSegments, 
                 null,
-                new MonitoringEventsWithSerilog(Logger.None))
+                new MonitoringEventsWithSerilog(Logger.None),
+                null)
         {
             var onRouteState = new OnRouteState(1, 2, TrackPoint.Unknown, DefaultSegments[0], Model.Route, SegmentDirection.AtoB, 0, 0, 0);
             Model.Route.EnteredSegment("seg-1");