+++ test/RoadCaptain.Tests.Unit/TestUserPreferences.cs
@@ -20,6 +20,8 @@ namespace RoadCaptain.Tests.Unit
         public bool ShowSprints { get; set; }
         public bool ShowClimbs { get; set; }
         public bool ShowElevationPlot { get; set; }
+        public CapturedWindowLocation? ElevationPlotWindowLocation { get; set; }
+        public bool ShowElevationPlotInGame { get; set; }
 
         public void Load()
         {